Crawler now receives a POST request and returns JSON
