+++ FlaskApp/crawlerUI/templates/crawl.html
@@ -7,7 +7,7 @@
 
 <!-- User Input Form -->
 <div class="panel-body">
-	<form action="crawl_submit" method="post" accept-charset="utf-8" class="form-horizontal">
+	<form action="crawler/" method="post" accept-charset="utf-8" class="form-horizontal">
 	<fieldset>
 		<legend>Crawl Web</legend>
